fix: correct emacs link check

+++ scripts/emacs.sh
@@ -29,7 +29,9 @@ main_brew() {
 	msg 'installation takes time and cpu because it compiles emacs from source'
 	brew install emacs-plus@29 --with-native-comp --with-elrumo2-icon
 
-	ln -s "$(brew --prefix)/opt/emacs-plus@29/Emacs.app" /Applications
+	if [ ! -e /Applications/Emacs.app ]; then
+		ln -s "$(brew --prefix)/opt/emacs-plus@29/Emacs.app" /Applications
+	fi
 
 	msg 'installing doom on osx manually is better because there are many errors'
 }
